+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 **Technical**
 - Remove support for Android 6
 - Target Android 17
+- Add pollen levels to datasharing lib
 
 
 # Version 6.2.2 (2026-08-23)